Use quint32 instead of unsigned long int as the return type for idleTime
unsigned long int doesn't work so well with QVariant, so let's use something
that's more compatible. :)

svn path=/branches/work/kopete/break-the-contactlist/; revision=707512

+++ libkopete/kopetemetacontact.h
@@ -348,7 +348,7 @@ public:
 	/**
 	 * return the time in second the contact is idle.
 	 */
-	unsigned long int idleTime() const;
+	quint32 idleTime() const;
 
 	/**
 	 * Get or set a field for the KDE address book backend. Fields not
